minimal trainable version

+++ extractor/data/datasets.py
@@ -88,6 +88,6 @@ class ScoreData(data.Dataset):
 
         # indexing numpy array with [None] adds extra channel dimension
         patch_score = torch.from_numpy(self.tilers[tomo].get_tile(self.score_volumes[tomo], idx)[None])
-        patch_label = torch.from_numpy(self.tilers[tomo].get_tile(self.label_volumes[tomo], idx)[None])
+        patch_label = torch.from_numpy(self.tilers[tomo].get_tile(self.label_volumes[tomo], idx)[None]).long()
 
         return patch_score, patch_label
